chore: cr feedback
