+++ core-libs/setup/ssr/providers/ssr-providers.ts
@@ -1,7 +1,8 @@
 import { StaticProvider } from '@angular/core';
 import { REQUEST } from '@nguniversal/express-engine/tokens';
 import { SERVER_REQUEST_ORIGIN, SERVER_REQUEST_URL } from '@spartacus/core';
-import { Request } from 'express';
+import { getRequestOrigin } from '../util/request-origin';
+import { getRequestUrl } from '../util/request-url';
 
 /**
  * Returns Spartacus providers to be passed to the Angular express engine (in SSR)
@@ -21,27 +22,4 @@ export function getServerRequestProviders(): StaticProvider[] {
       deps: [REQUEST],
     },
   ];
-}
-
-export function getRequestUrl(req: Request): string {
-  return getRequestOrigin(req) + req.originalUrl;
-}
-
-export function getRequestOrigin(req: Request): string {
-  // If express is resolving and trusting X-Forwarded-Host, we want to take it
-  // into an account to properly generate request origin.
-  const trustProxyFn = req.app.get('trust proxy fn');
-  let forwardedHost = req.get('X-Forwarded-Host');
-  if (forwardedHost && trustProxyFn(req.connection.remoteAddress, 0)) {
-    if (forwardedHost.indexOf(',') !== -1) {
-      // Note: X-Forwarded-Host is normally only ever a
-      //       single value, but this is to be safe.
-      forwardedHost = forwardedHost
-        .substring(0, forwardedHost.indexOf(','))
-        .trimRight();
-    }
-    return req.protocol + '://' + forwardedHost;
-  } else {
-    return req.protocol + '://' + req.get('host');
-  }
 }
